Integration of FDTD with SROPEE
Description:
The OIDT is a computational electromagnetic program. It is split into two portions (1) the Finite-Difference Time-Domain (FDTD) and scattering parameter extraction portion, both of which are located in the `.\FDTD` folder and (2) the equivalent circuit synthesis and passivity enforcement portion located in the `.\SROPEE` folder. A more detailed description is contained within the corresponding folders for each portion. Instructions for running the program from start to finish are in the `.\Integration_FDTD+SROPEE`.